refactor: consolidate reused code
