+++ lib/widgets/charts/time_series/time_series_line_chart.dart
@@ -6,42 +6,8 @@ import 'package:flutter/material.dart';
 import 'package:lotti/themes/colors.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/utils/platform.dart';
+import 'package:lotti/widgets/charts/time_series/utils.dart';
 import 'package:lotti/widgets/charts/utils.dart';
-
-const gridOpacity = 0.3;
-const labelOpacity = 0.5;
-
-class ChartLabel extends StatelessWidget {
-  const ChartLabel(this.text, {super.key});
-
-  final String text;
-
-  @override
-  Widget build(BuildContext context) {
-    return Opacity(
-      opacity: labelOpacity,
-      child: Text(
-        text,
-        style: chartTitleStyleSmall,
-        textAlign: TextAlign.center,
-      ),
-    );
-  }
-}
-
-List<Color> gradientColors = [
-  oldPrimaryColorLight,
-  oldPrimaryColor,
-];
-
-Widget leftTitleWidgets(double value, TitleMeta meta) {
-  return ChartLabel(value.toInt().toString());
-}
-
-final gridLine = FlLine(
-  color: chartTextColor.withOpacity(gridOpacity),
-  strokeWidth: 1,
-);
 
 class TimeSeriesLineChart extends StatelessWidget {
   const TimeSeriesLineChart({
